requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-datatable==0.3.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
MarkupSafe==2.1.1
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
virtualenv==20.16.3
Werkzeug==2.2.2
Flask==2.2.2
Jinja2==3.1.2
django-rest-auth==0.9.5
pip~=22.2.2
wheel~=0.37.1
setuptools==65.3.0
authentication~=1.1.0
helpers~=0.2.0
boto3==1.24.61
cryptography~=37.0.4
knox~=0.1.14
django-rest-knox==4.2.0
automium_web==0.1.1
